@@ -1,28 +1,25 @@
-using System;
 using System.Diagnostics;
-using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
-namespace CasCap.GooglePhotosCli.Tests
-{
-    public class CliTests : TestBase
-    {
-        public CliTests(ITestOutputHelper output) : base(output) { }
+namespace CasCap.GooglePhotosCli.Tests;
 
-        [Fact]
-        public async Task Test1()
+public class CliTests : TestBase
+{
+    public CliTests(ITestOutputHelper output) : base(output) { }
+
+    [Fact]
+    public async Task Test1()
+    {
+        //todo: how best to test a command line application?
+        try
         {
-            //todo: how best to test a command line application?
-            try
-            {
-                _ = await _googlePhotosSvc.CreateAlbumAsync("test");
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine(ex);
-                Debugger.Break();
-            }
-            Assert.True(true);//assert true regardless of actual outcome, will add full tests later
+            _ = await _googlePhotosSvc.CreateAlbumAsync("test");
         }
+        catch (Exception ex)
+        {
+            Debug.WriteLine(ex);
+            Debugger.Break();
+        }
+        Assert.True(true);//assert true regardless of actual outcome, will add full tests later
     }
 }
