@@ -1,20 +1,18 @@
 ﻿using CasCap.Services;
 using McMaster.Extensions.CommandLineUtils;
-using System.Threading.Tasks;
-namespace CasCap.Commands
+namespace CasCap.Commands;
+
+[Command(Description = "Sign-out and delete all local data.")]
+internal class Logout : CommandBase
 {
-    [Command(Description = "Sign-out and delete all local data.")]
-    internal class Logout : CommandBase
+    public Logout(IConsole console, DiskCacheService diskCacheSvc, GooglePhotosService googlePhotosSvc) : base(console, diskCacheSvc, googlePhotosSvc) { }
+
+    public async override Task<int> OnExecuteAsync(CommandLineApplication app)
     {
-        public Logout(IConsole console, DiskCacheService diskCacheSvc, GooglePhotosService googlePhotosSvc) : base(console, diskCacheSvc, googlePhotosSvc) { }
+        await base.OnExecuteAsync(app);
 
-        public async override Task<int> OnExecuteAsync(CommandLineApplication app)
-        {
-            await base.OnExecuteAsync(app);
+        _console.WriteLine($"todo: need to implement this...");
 
-            _console.WriteLine($"todo: need to implement this...");
-
-            return 0;
-        }
+        return 0;
     }
 }
