@@ -1,9 +1,7 @@
-﻿using System;
-namespace CasCap.Models
+﻿namespace CasCap.Models;
+
+public class AppConfig
 {
-    public class AppConfig
-    {
-        public DateTime lastCheck { get; set; } = DateTime.MinValue;
-        public DateTime latestMediaItemCreation { get; set; } = DateTime.MinValue;
-    }
+    public DateTime lastCheck { get; set; } = DateTime.MinValue;
+    public DateTime latestMediaItemCreation { get; set; } = DateTime.MinValue;
 }
