@@ -1,11 +1,10 @@
-﻿namespace CasCap.ViewModels
+﻿namespace CasCap.ViewModels;
+
+public struct MediaItemScore
 {
-    public struct MediaItemScore
-    {
-        public int count { get; set; }
+    public int count { get; set; }
 
-        public GroupByProperty propertyMatches { get; set; }
+    public GroupByProperty propertyMatches { get; set; }
 
-        public override string ToString() => $"{count} - {propertyMatches}";
-    }
+    public override string ToString() => $"{count} - {propertyMatches}";
 }
